Add driver for Logitech M560
The Logitech M560 is is a mouse designed for windows 8.
Comparing to a standard one, some buttons (the middle one and the
two ones placed on the side) are bounded to a key combination
instead of a classic "mouse" button.

Think this mouse as a pair of mouse and keyboard. When the middle
button is pressed the it sends a key (as keyboard)
combination, the same for the other two side button.
Instead the left/right/wheel work correctly.
To complicate further the things, the middle button send a
key combination the odd press, and another one for the even press;
in the latter case it sends also a left click. But the worst thing
is that no event is generated when the middle button is released.

Moreover this device is a wireless mouse which uses the unifying
receiver.

It is possible to re-configure the mouse
sending a command (see function m560_send_config_command()).
After this command the mouse sends some sequence when the
buttons are pressed and/or released (see comments for
an explanation of the mouse protocol).

Signed-off-by: Goffredo Baroncelli <kreijack@inwind.it>